kernel/int: consolidate interrupt vector number to `int32`
This is no-op for 32 bit platforms because `int32` is defined as `long` there.

Change interrupt vector number from 64 bits to 32 bits for 64 bit platforms.

Reviewed-on: https://review.haiku-os.org/c/haiku/+/7507
Tested-by: Commit checker robot <no-reply+buildbot@haiku-os.org>

+++ headers/os/drivers/KernelExport.h
@@ -168,9 +168,9 @@ extern void			release_write_seqlock(seqlock* lock);
 extern uint32		acquire_read_seqlock(seqlock* lock);
 extern bool			release_read_seqlock(seqlock* lock, uint32 count);
 
-extern status_t		install_io_interrupt_handler(long interrupt_number,
-						interrupt_handler handler, void *data, ulong flags);
-extern status_t		remove_io_interrupt_handler(long interrupt_number,
+extern status_t		install_io_interrupt_handler(int32 interrupt_number,
+						interrupt_handler handler, void *data, uint32 flags);
+extern status_t		remove_io_interrupt_handler(int32 interrupt_number,
 						interrupt_handler handler, void	*data);
 
 extern status_t		add_timer(timer *t, timer_hook hook, bigtime_t period,
